% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UDID-package.R
\docType{package}
\name{UDID-package}
\alias{UDID}
\alias{UDID-package}
\title{UDID: Universal Difference-in-Differences Estimator}
\description{
Parametric and nonparametric universal difference-in-differences
estimators with sensitivity analysis (Park & Tchetgen Tchetgen, 2025).
}
\author{
\strong{Maintainer}: Chan Park \email{parkchan@illinois.com} (Park & Tchetgen Tchetgen, 2025)

}
